Have AST.edges use HashSet instead of Vector.
AST.edges now is a HashSet instead of being a Vector. This is an
internal change for the most part. I'm using HashSet because the
property of 'no duplication' is important for the edges list, however
the exact order isn't important. AST.nodes is still a Vector as the
order is important more than anything. Duplicate nodes don't necessarily
matter as much.